Add MOVE command
It was a bit difficult to decide on how to handle how to move the
Robot instance itself, whether to update its coordinates per OOP
or go FP and create a new Robot with the modified coordinate.
I decided on the stateful OOP route, because it felt more natural,
and I figured I could have a little statefulness, as a treat.