Add LEFT command
As with the implementation of MOVE, I decided to make the Robot
model stateful to handle any changes in its position/orientation.

I also updated RobotFactory to use TableFactory.default, because
managing table instances in specs was becoming a pain. I'll probably
want to refactor Table functionality a bit soon in order to separate
it better from Robot.