Add Table persistence
Creating a new table every time we wanted to create a new Robot,
or fetch one from the data store works okay when we have a fixed
table size, but it's a brittle design and will break if we add
variable table sizes. So, it makes sense to separate the Table
lifecycle from that of the Robot and persist the attributes of the
former for easy retrieval.