Remove TableRepository dependency from RobotRepository
This was a temporary measure to break up the changes to the
RobotService a bit, since we needed a table to create a valid
Robot, and returning an attributes hash instead of a Robot instance
broke a bunch of expectations.

Table and Robot logic are now less coupled, with object management
relegated to the RobotService class.

I also added more logic to the factories to simplify test setup.